pyproject.toml
[project]
name = "suezcanal"
version = "1.0.0"
description = "SuezCanal.xyz — Open-source maritime SAR and threat awareness platform"
readme = "README.md"
license = { text = "AGPL-3.0-or-later" }
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",
    "pydantic>=2.7",
    "pydantic-settings>=2.2",
    "sqlalchemy[asyncio]>=2.0",
    "aiosqlite>=0.20",
    "alembic>=1.13",
    "httpx>=0.27",
    "websockets>=12.0",
    "redis>=5.0",
    "pynacl>=1.5",
    "blake3>=0.3",
    "click>=8.1",
    "rich>=13.7",
    "python-dateutil>=2.9",
    "python-dotenv>=1.0",
    "requests>=2.31",
    "copernicusmarine>=2.3.0",
    "xarray>=2025.0.0",
]

[project.optional-dependencies]
sdr = ["pyrtlsdr>=0.3"]
tid = [
    "missile-tid @ git+https://github.com/tylerni7/missile-tid.git",
    "cartopy>=0.22",
    "shapely>=2.0",
    "pycurl>=7.45",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "mypy>=1.10",
]
full = ["suezcanal[sdr,tid,dev]"]

[build-system]
requires = ["setuptools>=70"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["core*", "sdk*", "edge*", "connect*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
